// src/types/exoplanets/habitable.types.ts
export type HabitableWorldDetails = {
    pl_name: string | null;
    hostname: string | null;
    sy_dist: number | null;
    disc_year: number | null;
    discoverymethod: string | null;
    pl_rade: number | null;
    pl_bmasse: number | null;
    pl_orbper: number | null;
    pl_eqt: number | null;
    st_teff: number | null;
    st_rad: number | null;
    st_mass: number | null;
    st_age: number | null;
    st_spectype: string | null;
    habitabilityScore: number;
    habitabilityZone: string;
};